split pregnancy script
